made hook for logout we just need to simply import the function handleLogout
